Move ai handlers to specific folder
