+++ pr_agent/algo/ai_handlers/litellm_ai_handler.py
@@ -7,7 +7,7 @@ from openai.error import APIError, RateLimitError, Timeout, TryAgain
 from retry import retry
 from pr_agent.config_loader import get_settings
 from pr_agent.log import get_logger
-from pr_agent.algo.base_ai_handler import BaseAiHandler
+from pr_agent.algo.ai_handlers.base_ai_handler import BaseAiHandler
 OPENAI_RETRIES = 5
