+++ pr_agent/algo/ai_handlers/openai_ai_handler.py
@@ -1,4 +1,4 @@
-from pr_agent.algo.base_ai_handler import BaseAiHandler
+from pr_agent.algo.ai_handlers.base_ai_handler import BaseAiHandler
 import openai
 from openai.error import APIError, RateLimitError, Timeout, TryAgain
 from retry import retry
