+++ pr_agent/tools/pr_code_suggestions.py
@@ -4,7 +4,7 @@ from typing import Dict, List
 
 from jinja2 import Environment, StrictUndefined
 
-from pr_agent.algo.base_ai_handler import BaseAiHandler
+from pr_agent.algo.ai_handlers.base_ai_handler import BaseAiHandler
 from pr_agent.algo.pr_processing import get_pr_diff, get_pr_multi_diffs, retry_with_fallback_models
 from pr_agent.algo.token_handler import TokenHandler
 from pr_agent.algo.utils import load_yaml, get_ai_handler
