+++ pr_agent/tools/pr_update_changelog.py
@@ -5,7 +5,7 @@ from typing import Tuple
 
 from jinja2 import Environment, StrictUndefined
 
-from pr_agent.algo.base_ai_handler import BaseAiHandler
+from pr_agent.algo.ai_handlers.base_ai_handler import BaseAiHandler
 from pr_agent.algo.pr_processing import get_pr_diff, retry_with_fallback_models
 from pr_agent.algo.token_handler import TokenHandler
 from pr_agent.algo.utils import get_ai_handler
